# Improving "Sparse postings" intersection