Lets take the given example:

P1: [2, 5, 9, 18, 21]
P2: [3, 7, 14, 19, 21]
P3: [1, 21]

Currently, we would only advance through P1 and P2 until discovering an intersection and then checking P3. In essence, the traversal order was: 2, 3, 5, 7, 9, 14, 18, 19, 21 (intersection found).

With the proposed change, P3 is also examined even if P1 and P2 haven't found an intersection yet. This adjustment allows for the possibility of skipping some iterations.

Post-change, the traversal order becomes: 2, 3, 21 (3 iterations instead of 9).

To validate the improvement, benchmarks were adjusted to simulate this scenario. Additionally, calling next on the resulting postings demonstrates the benefits. In this extreme case, a significant 97% reduction in time is observed.

```

goos: linux
goarch: amd64
pkg: github.com/prometheus/prometheus/tsdb
cpu: Intel(R) Xeon(R) Platinum 8259CL CPU @ 2.50GHz
                                                          │   /tmp/old    │              /tmp/new              │
                                                          │    sec/op     │   sec/op     vs base               │
Querier/Block/PostingsForMatchers/n="0_1",j="foo,a="1"-32   15815.1µ ± 7%   431.4µ ± 3%  -97.27% (p=0.002 n=6)

                                                          │  /tmp/old  │           /tmp/new            │
                                                          │    B/op    │    B/op     vs base           │
Querier/Block/PostingsForMatchers/n="0_1",j="foo,a="1"-32   336.0 ± 0%   336.0 ± 0%  ~ (p=1.000 n=6) ¹
¹ all samples are equal

                                                          │  /tmp/old  │           /tmp/new            │
                                                          │ allocs/op  │ allocs/op   vs base           │
Querier/Block/PostingsForMatchers/n="0_1",j="foo,a="1"-32   13.00 ± 0%   13.00 ± 0%  ~ (p=1.000 n=6) ¹
¹ all samples are equal
```

Ideally we could  sort the `arr []Postings` by size before iterating but, unfortunately,  the postings interface does not allow us to retrieve the underlining size.