@GiedriusS 2024-04-23
I tried adding a length method here https://github.com/prometheus/prometheus/pull/13093, maybe worth revisiting?