@alanprot 2024-07-04
@bboreham @jesusvazquez Any thoughts on this?

This is can be pretty bad if we have a query where we have small and large postings:

``
max(cpu_utilization{env="prod", service="a"})
``

Where `cpu_utilization` and `env="prod"` are large postings but `service=a` is a small one.