@bboreham 2024-07-05
Sorry, I had not previously noticed this PR. 

I found the code quite difficult to understand.
The idea is to seek all iterators until they are past the one being compared?  So that should improve sparse cases, but will it worsen dense matches?

Perhaps naming could be improved to make the behaviour more understandable.