@alanprot 2024-09-12
> The idea is to seek all iterators until they are past the one being compared? So that should improve sparse cases, but will it worsen dense matches?

The key is to ensure that small postings aren't overlooked and to give every posting an equal opportunity. If we process all postings at the same pace, we avoid favoring the larger ones, which could lead to missing out on quicker results from the smaller postings.

If we knew the size of the postings beforehand we could sort to start iterating from the smaller ones... this would be ideal i think.